requirements.txt
tensorflow
numpy
azureml-core
